backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.19

psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
cryptography==44.0.0
httpx==0.27.2
